simd scalar emu compiles but fails in pq4 search
